・GameManagerからViewに分割 ・constantファイル作成 ・Creatureファイル作成
