--- Unity/Assets/Animation/C1BattleAnimation1.anim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: e6581ef8ec08b614d82c24b5cb946f2c
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 7400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
