--- Unity/Assets/Animation/C1BattleAnimation2.anim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 314e8bcfc56df604e9522d271171b57c
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 7400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
