--- Unity/Assets/Animation/C2BattleAnimation1.anim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 23b3410de71c303408f0e8eb703f8d41
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 7400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
