--- Unity/Assets/Animation/C2BattleAnimation2.anim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ce3b2c87c6d096247994a7c0e38ec288
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 7400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
