--- Unity/Assets/Animation/CreatureBattleImage1.controller.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 402ec5fda9e637345ba3b673b8129ee0
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 9100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
