--- Unity/Assets/Animation/CreatureBattleImage2.controller.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f5a5f79225a411c4ea2834e4c85bc39f
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 9100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
