--- Unity/Assets/Prefabs.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 55faa5712528b9a4ba527b97d3a677c7
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
