--- Unity/Assets/Prefabs/煙.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 227b1beb3b2af1f479f85841e08a718a
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
