--- Unity/Assets/Scripts/ChangeFieldImageModel.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: fa52e7a4889cbe041b04299b3e2258eb
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
