+++ Unity/Assets/Scripts/Creature.cs
@@ -1,0 +1,13 @@
+public class Creature
+{
+    public string name;
+    public int power;
+    public string field;
+
+    public Creature(string name, int power, string field)
+    {
+        this.name = name;
+        this.power = power;
+        this.field = field;
+    }
+}
