+++ Unity/Assets/Scripts/Creature.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 52edb5852fe9e614e87cbff16383b607
+guid: 047d999d07c5c7343864236bbbcb127e
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
