+++ Unity/Assets/Scripts/CreatureConstant.cs
@@ -1,0 +1,12 @@
+public class CreatureConstant
+{
+    public enum Creature
+    {
+        Elephant = 0,
+        Lion = 1,
+        Zebra = 2,
+        Dolphin = 3,
+        Orca = 4,
+        Human = 5,
+    }
+}
