+++ Unity/Assets/Scripts/CreatureConstant.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: ceda70b7872a4a940b55acaf2a69e391
+guid: afcff352e46fa714c9d16f9b801e28e0
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
