+++ Unity/Assets/Scripts/CreatureShowView.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: fcbf828b39be5ed4e84c66d66aabef85
+guid: 700f1a03adf39e441b14708818b04d29
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
