--- Unity/Assets/Scripts/DeleteButton1Script.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 93050edb66a57ba4db17869917096a8e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
