--- Unity/Assets/Scripts/ElephantButtonScript.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9611c972ff7ee34418d04db889434241
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
