+++ Unity/Assets/Scripts/FieldConstant.cs
@@ -1,0 +1,15 @@
+public class FieldConstant
+{
+    public const string Tokyo = "東京";
+    public const string Savannah = "サバンナ";
+    public const string PacificOcean = "太平洋";
+    public const string Space = "宇宙";
+
+    public enum Field
+    {
+        Tokyo = 0,
+        Savannah = 1,
+        PacificOcean = 2,
+        Space = 3,
+    }
+}
