+++ Unity/Assets/Scripts/FieldConstant.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: da89bae990a83734395f511f7610bb63
+guid: 8ac007a60842e5846ad98af095e4a141
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
