--- Unity/Assets/Scripts/FinishButtonScript.cs
@@ -1,22 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class FinishButtonScript : MonoBehaviour
-{
-    // Start is called before the first frame update
-    void Start()
-    {
-        
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-
-    }
-
-    public void OnClick(){
-        Application.Quit();
-    }
-}
