--- Unity/Assets/Scripts/NextGameButtonScript.cs
@@ -1,24 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.SceneManagement;
-
-
-public class NextGameButtonScript : MonoBehaviour
-{
-    // Start is called before the first frame update
-    void Start()
-    {
-        
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        
-    }
-
-    public void OnClick(){
-        SceneManager.LoadScene("MainScene");
-    }
-}
