--- Unity/Assets/Scripts/NextGameButtonScript.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 33dfc2c273ec688428a975d6dbbc9db5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
