--- Unity/Assets/Scripts/OrcaButtonScript.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b46a75a3a71f8f94d901cf5f49fb7086
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
