--- Unity/Assets/Scripts/ReturnButtonScript.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ab76b11bfc684544994868c90dbcef85
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
